better architecture and naming of packages and files
