+++ generate_test.go
@@ -2,8 +2,8 @@ package identity
 
 import (
 	"encoding/base64"
+	"github.com/TankerHQ/identity-go/v3/internal/base64_json"
 
-	"github.com/TankerHQ/identity-go/v3/b64json"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"golang.org/x/crypto/blake2b"
@@ -36,7 +36,7 @@ var _ = Describe("generate", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		id := &identity{}
-		err = b64json.Decode(*identityB64, id)
+		err = base64_json.Decode(*identityB64, id)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(id.Target).Should(Equal("user"))
 	})
@@ -45,10 +45,10 @@ var _ = Describe("generate", func() {
 		goodPermanentIdentity := "eyJ0cnVzdGNoYWluX2lkIjoidHBveHlOemgwaFU5RzJpOWFnTXZIeXlkK3BPNnpHQ2pPOUJmaHJDTGpkND0iLCJ0YXJnZXQiOiJ1c2VyIiwidmFsdWUiOiJSRGEwZXE0WE51ajV0VjdoZGFwak94aG1oZVRoNFFCRE5weTRTdnk5WG9rPSIsImRlbGVnYXRpb25fc2lnbmF0dXJlIjoiVTlXUW9sQ3ZSeWpUOG9SMlBRbWQxV1hOQ2kwcW1MMTJoTnJ0R2FiWVJFV2lyeTUya1d4MUFnWXprTHhINmdwbzNNaUE5cisremhubW9ZZEVKMCtKQ3c9PSIsImVwaGVtZXJhbF9wdWJsaWNfc2lnbmF0dXJlX2tleSI6IlhoM2kweERUcHIzSFh0QjJRNTE3UUt2M2F6TnpYTExYTWRKRFRTSDRiZDQ9IiwiZXBoZW1lcmFsX3ByaXZhdGVfc2lnbmF0dXJlX2tleSI6ImpFRFQ0d1FDYzFERndvZFhOUEhGQ2xuZFRQbkZ1Rm1YaEJ0K2lzS1U0WnBlSGVMVEVOT212Y2RlMEhaRG5YdEFxL2RyTTNOY3N0Y3gwa05OSWZodDNnPT0iLCJ1c2VyX3NlY3JldCI6IjdGU2YvbjBlNzZRVDNzMERrdmV0UlZWSmhYWkdFak94ajVFV0FGZXh2akk9In0="
 
 		id := &identity{}
-		err := b64json.Decode(goodPermanentIdentity, id)
+		err := base64_json.Decode(goodPermanentIdentity, id)
 		Expect(err).ShouldNot(HaveOccurred())
 
-		orderedJson, _ := b64json.Encode(id)
+		orderedJson, _ := base64_json.Encode(id)
 
 		Expect(goodPermanentIdentity).Should(Equal(*orderedJson))
 	})
@@ -116,7 +116,7 @@ var _ = Describe("generate", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		id := &provisionalIdentity{}
-		err = b64json.Decode(*identityB64, id)
+		err = base64_json.Decode(*identityB64, id)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(id.Target).Should(Equal("email"))
 		Expect(id.Value).Should(Equal("email@example.com"))
@@ -131,7 +131,7 @@ var _ = Describe("generate", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		id := &provisionalIdentity{}
-		err = b64json.Decode(*identityB64, id)
+		err = base64_json.Decode(*identityB64, id)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(id.Target).Should(Equal("phone_number"))
 		Expect(id.Value).Should(Equal(userPhone))
@@ -143,7 +143,7 @@ var _ = Describe("generate", func() {
 
 	It("deserializes a valid phone_number provisional identity", func() {
 		id := &provisionalIdentity{}
-		err := b64json.Decode(phoneNumberProvisionalIdentity, id)
+		err := base64_json.Decode(phoneNumberProvisionalIdentity, id)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(id.Target).Should(Equal("phone_number"))
 		Expect(id.Value).Should(Equal(userPhone))
@@ -159,7 +159,7 @@ var _ = Describe("generate", func() {
 
 	It("deserializes a valid phone_number public provisional identity", func() {
 		privId := &provisionalIdentity{}
-		err := b64json.Decode(phoneNumberProvisionalIdentity, privId)
+		err := base64_json.Decode(phoneNumberProvisionalIdentity, privId)
 		Expect(err).ShouldNot(HaveOccurred())
 		hashedPhone := hashProvisionalIdentityValue(userPhone, base64.StdEncoding.EncodeToString(privId.PrivateSignatureKey))
 
@@ -167,7 +167,7 @@ var _ = Describe("generate", func() {
 		Expect(*publicIdentity).Should(Equal(phoneNumberPublicProvisionalIdentity))
 
 		id := &publicProvisionalIdentity{}
-		err2 := b64json.Decode(*publicIdentity, id)
+		err2 := base64_json.Decode(*publicIdentity, id)
 		Expect(err2).ShouldNot(HaveOccurred())
 		Expect(id.Target).Should(Equal("hashed_phone_number"))
 		Expect(id.Value).Should(Equal(hashedPhone))
@@ -182,10 +182,10 @@ var _ = Describe("generate", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		extractedPublicID := &publicIdentity{}
-		_ = b64json.Decode(*publicID, extractedPublicID)
+		_ = base64_json.Decode(*publicID, extractedPublicID)
 
 		extractedGoodPublicID := &publicIdentity{}
-		_ = b64json.Decode(goodPublicIdentity, extractedGoodPublicID)
+		_ = base64_json.Decode(goodPublicIdentity, extractedGoodPublicID)
 		Expect(*extractedGoodPublicID).Should(Equal(*extractedPublicID))
 	})
 
@@ -196,7 +196,7 @@ var _ = Describe("generate", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		provisionalID := &provisionalIdentity{}
-		_ = b64json.Decode(*identityB64, provisionalID)
+		_ = base64_json.Decode(*identityB64, provisionalID)
 		Expect(err).ShouldNot(HaveOccurred())
 
 		publicID, err := GetPublicIdentity(*identityB64)
@@ -207,7 +207,7 @@ var _ = Describe("generate", func() {
 		expectedId.Value = base64.StdEncoding.EncodeToString(hashedEmail[:])
 
 		extractedPublicID := &publicProvisionalIdentity{}
-		_ = b64json.Decode(*publicID, extractedPublicID)
+		_ = base64_json.Decode(*publicID, extractedPublicID)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(*extractedPublicID).Should(Equal(expectedId))
 	})
