+++ identity/go.mod
@@ -1,0 +1,13 @@
+module github.com/TankerHQ/identity-go/v3/identity
+
+go 1.17
+
+require github.com/TankerHQ/identity-go/v3 v3.1.1-0.20210915140450-415f3543a979
+
+require (
+	github.com/iancoleman/orderedmap v0.2.0 // indirect
+	github.com/onsi/ginkgo v1.16.4 // indirect
+	github.com/onsi/gomega v1.16.0 // indirect
+	golang.org/x/crypto v0.0.0-20210915214749-c084706c2272 // indirect
+	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+)
