+++ internal/crypto/crypto_test.go
@@ -1,14 +1,14 @@
-package curve25519_test
+package crypto_test
 
 import (
-	"github.com/TankerHQ/identity-go/v3/curve25519"
+	"github.com/TankerHQ/identity-go/v3/internal/crypto"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("curve25519Generate", func() {
 	It("generates a non nil key pair", func() {
-		sk, pk, err := curve25519.GenerateKey()
+		sk, pk, err := crypto.GenerateKey()
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(len(sk)).Should(Equal(32))
 		Expect(len(pk)).Should(Equal(32))
@@ -19,9 +19,9 @@ var _ = Describe("curve25519Generate", func() {
 	})
 
 	It("generates different keys each time", func() {
-		sk, pk, err := curve25519.GenerateKey()
+		sk, pk, err := crypto.GenerateKey()
 		Expect(err).ShouldNot(HaveOccurred())
-		sk2, pk2, err := curve25519.GenerateKey()
+		sk2, pk2, err := crypto.GenerateKey()
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(sk).ShouldNot(Equal(sk2))
 		Expect(pk).ShouldNot(Equal(pk2))
