+++ user_secret_test.go
